Disable Bluetooth support
The integrated browser is not supposed to be a fully featured web
browser, but rather provide support for modern web standards that can
be integrated neatly into an overlay or the UI via a panel.

WebBluetooth requires additional system permissions (especially on
macOS) and can potentially cause OBS to crash if those permissions are
denied. That's not ideal either.